feat: modify data layer frame
- change service naming to datasource
- add remoteDataSource to wrap retrofit service
- remove repository interface to fit the Google app architecture
