+++ app/src/main/java/com/lyrics/feelin/core/data/datasource/remote/AuthApiService.kt
@@ -1,0 +1,3 @@
+package com.lyrics.feelin.core.data.datasource.remote
+
+interface AuthApiService
