+++ app/src/main/java/com/lyrics/feelin/core/data/datasource/remote/AuthRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.lyrics.feelin.core.data.datasource.remote
+
+// TODO(@이대근): 어노테이션은 구현 중 제거할 것. 2025.10.02.
+@Suppress("UnusedPrivateProperty")
+class AuthRemoteDataSource(
+    private val authApiService: AuthApiService
+)
