--- app/src/main/java/com/lyrics/feelin/core/data/service/AuthApiService.kt
@@ -1,3 +1,0 @@
-package com.lyrics.feelin.core.data.service
-
-interface AuthApiService
