--- app/src/main/java/com/lyrics/feelin/core/data/service/KakaoOAuthSdkService.kt
@@ -1,7 +1,0 @@
-package com.lyrics.feelin.core.data.service
-
-class KakaoOAuthSdkService {
-    constructor() {
-        TODO("Not yet implemented")
-    }
-}
